Fixed provider abstraction, split services.
