--- ferv_project/recommendation/versions/V1/urls.py
@@ -1,8 +1,0 @@
-from django.urls import path
-from recommendation.views import RecommendationView
-
-app_name = 'recommendation'
-
-urlpatterns = [
-    path('', RecommendationView.as_view(), name='recommend'),
-]
